requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
fastapi==0.134.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
future==1.0.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
line-bot-sdk==3.22.0
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numpy==2.4.1
pillow==12.1.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
setuptools==80.10.2
six==1.17.0
starlette==0.52.1
sympy==1.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
wrapt==2.1.1
yarl==1.22.0
